refactor: meta data constants
