+++ utils/constants/constants.ts
@@ -1,33 +1,11 @@
-import type { Metadata } from "next";
 export const IS_PROD = process.env.NODE_ENV === "production" ? true : false;
 export const BASE_URL = IS_PROD
   ? "https://community.techtankto.com"
   : "http://localhost:3000";
 
-const title_template = "TechTank Community Portal";
-const description_template =
+export const TITLE_TEMPLATE = "TechTank Community Portal";
+export const DESCRIPTION_TEMPLATE =
   "Welcome to the TechTank community portal, where you will find important information around ";
 
-export const generateMetadataObject = (
-  title = title_template,
-  description = description_template
-): Metadata => {
-  const metadata: Metadata = {
-    title,
-    description,
-    keywords: ["TechTank", "Toronto", "JavaScript", "Tech community", "Tech"],
-    openGraph: {
-      title,
-      description,
-      url: BASE_URL,
-      siteName: "TechTank Community Portal",
-    },
-    twitter: {
-      card: "summary_large_image",
-      title,
-      description,
-    },
-  };
-
-  return metadata;
-};
+export const SEO_IMG =
+  "https://res.cloudinary.com/dhu6qys0s/image/upload/v1728506219/community_seo_img_fahysm.png";
