+++ utils/constants/index.ts
@@ -1,3 +1,3 @@
 export { rubik } from "./fonts";
 export { IS_PROD } from "./constants";
-export { generateMetadataObject } from "./constants";
+export { generateMetadataObject } from "./site_metadata";
